add set people to App, props going down to People people is making the api call and sending the data back up app.  Setting apps state to the people information.
